move capsule commands to the workspace extension
